HW10: move logger & logger_uber to /internal/pkg. Rename logger_uber to loggerzap
