+++ hw12_13_14_15_calendar/internal/pkg/loggerzap/loggerzap.go
@@ -1,4 +1,4 @@
-package logger_uber
+package loggerzap
 
 import (
 	"fmt"
